(panel_save_setup): fix comparison between signed and unsigned.

+++ src/setup.c
@@ -1412,7 +1412,7 @@ panel_save_setup (WPanel * panel, const char *section)
     mc_config_set_string (mc_panels_config, section, "sort_order", panel->sort_field->id);
 
     for (i = 0; list_types[i].key != NULL; i++)
-        if (list_types[i].list_type == panel->list_type)
+        if (list_types[i].list_type == (int) panel->list_type)
         {
             mc_config_set_string (mc_panels_config, section, "list_mode", list_types[i].key);
             break;
